Added prevLine, AuthorSearchCache and moved search single page
If the last line of the previous page hadn't ended it is appended to the
first line of the next page and search with it.
Added a cache to hold variables about a single author's search to make
it easier to split up the code.
Split the code for searching a single page into it's own method.